@@ -1,7 +1,7 @@
 {
     "name": "chimera/di-symfony",
-    "description": "A set of compiler passes that connects all components together",
     "type": "library",
+    "description": "A set of compiler passes that connects all components together",
     "license": "MIT",
     "authors": [
         {
@@ -9,33 +9,40 @@
             "email": "lcobucci@gmail.com"
         }
     ],
-    "config": {
-        "preferred-install": "dist",
-        "sort-packages": true
-    },
     "require": {
-        "php": "^7.3 || 8.0",
-        "chimera/routing": "^0.3",
-        "lcobucci/di-builder": "^5.6"
+        "php": "^7.4 || ^8.0",
+        "chimera/foundation": "^0.4.x-dev@dev",
+        "chimera/routing": "^0.4.x-dev@dev",
+        "lcobucci/di-builder": "^7.0"
     },
     "require-dev": {
-        "chimera/bus-tactician": "^0.3",
-        "chimera/mapping": "^0.3",
-        "chimera/routing-expressive": "^0.3",
-        "chimera/serialization-jms": "^0.3",
-        "infection/infection": "^0.16",
-        "lcobucci/coding-standard": "^3.1",
-        "phpstan/phpstan": "^0.11",
-        "phpstan/phpstan-deprecation-rules": "^0.11",
-        "phpstan/phpstan-phpunit": "^0.11",
-        "phpstan/phpstan-strict-rules": "^0.11",
-        "phpunit/phpunit": "^9.1"
+        "chimera/bus-tactician": "^0.4.x-dev@dev",
+        "chimera/mapping": "^0.4.x-dev@dev",
+        "chimera/routing-mezzio": "^0.4.x-dev@dev",
+        "chimera/serialization-jms": "^0.4.x-dev@dev",
+        "infection/infection": "^0.20",
+        "lcobucci/coding-standard": "^6.0",
+        "phpstan/extension-installer": "^1.1",
+        "phpstan/phpstan": "^0.12",
+        "phpstan/phpstan-deprecation-rules": "^0.12",
+        "phpstan/phpstan-phpunit": "^0.12",
+        "phpstan/phpstan-strict-rules": "^0.12",
+        "phpunit/phpunit": "^9.5"
     },
     "suggest": {
         "chimera/bus-tactician": "To use league/tactician as service bus",
         "chimera/mapping": "To use annotations for mapping services",
-        "chimera/serialization-jms": "To use jms/serializer to create messages and format content",
-        "chimera/routing-expressive": "To use zendframework/zend-expressive as routing application"
+        "chimera/routing-mezzio": "To use mezzio/mezzio as routing application",
+        "chimera/serialization-jms": "To use jms/serializer to create messages and format content"
+    },
+    "config": {
+        "preferred-install": "dist",
+        "sort-packages": true
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "0.4.x-dev"
+        }
     },
     "autoload": {
         "psr-4": {
@@ -45,11 +52,6 @@
     "autoload-dev": {
         "psr-4": {
             "Chimera\\DependencyInjection\\Tests\\": "tests"
-        }
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "0.4.x-dev"
         }
     }
 }
