@@ -7,20 +7,15 @@ use Generator;
 use Lcobucci\DependencyInjection\CompilerPassListProvider;
 use Lcobucci\DependencyInjection\FileListProvider;
 use Symfony\Component\DependencyInjection\Compiler\PassConfig;
-use function assert;
+
 use function dirname;
 
 final class RegisterApplication implements FileListProvider, CompilerPassListProvider
 {
-    /**
-     * @var string
-     */
-    private $name;
+    private string $name;
 
-    /**
-     * @var ConditionallyLoadedPackage[]
-     */
-    private $relatedPackages;
+    /** @var list<ConditionallyLoadedPackage> */
+    private array $relatedPackages;
 
     public function __construct(string $name)
     {
@@ -33,7 +28,7 @@ final class RegisterApplication implements FileListProvider, CompilerPassListPro
             new MessageCreator\JmsSerializer\Package(),
             new Mapping\Package(),
             new ServiceBus\Tactician\Package($commandBusId, $queryBusId),
-            new Routing\Expressive\Package($name, $commandBusId, $queryBusId),
+            new Routing\Mezzio\Package($name, $commandBusId, $queryBusId),
         ];
     }
 
@@ -43,8 +38,6 @@ final class RegisterApplication implements FileListProvider, CompilerPassListPro
         yield dirname(__DIR__) . '/config/routing.xml';
 
         foreach ($this->filterPackages(FileListProvider::class) as $package) {
-            assert($package instanceof FileListProvider);
-
             yield from $package->getFiles();
         }
     }
@@ -55,18 +48,25 @@ final class RegisterApplication implements FileListProvider, CompilerPassListPro
         yield [new ValidateApplicationComponents($this->name), PassConfig::TYPE_OPTIMIZE, -30];
 
         foreach ($this->filterPackages(CompilerPassListProvider::class) as $package) {
-            assert($package instanceof CompilerPassListProvider);
-
             yield from $package->getCompilerPasses();
         }
     }
 
+    /**
+     * @template T
+     *
+     * @param class-string<T> $type
+     *
+     * @return Generator<T>
+     */
     private function filterPackages(string $type): Generator
     {
         foreach ($this->relatedPackages as $package) {
-            if ($package instanceof $type && $package->shouldBeLoaded()) {
-                yield $package;
+            if (! $package instanceof $type || ! $package->shouldBeLoaded()) {
+                continue;
             }
+
+            yield $package;
         }
     }
 }
