@@ -14,10 +14,13 @@ use Symfony\Component\Config\Resource\FileResource;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
+
 use function assert;
+use function class_exists;
 use function get_class;
 use function implode;
 use function is_array;
+use function is_string;
 
 final class ExpandTags implements CompilerPassInterface
 {
@@ -75,20 +78,25 @@ final class ExpandTags implements CompilerPassInterface
         $reader = Mapping\Reader::fromDefault();
 
         foreach ($container->getDefinitions() as $definition) {
-            $class       = new ReflectionClass($definition->getClass());
+            $className = $definition->getClass();
+            assert(is_string($className));
+            assert(class_exists($className));
+
+            $class       = new ReflectionClass($className);
             $annotations = $reader->getClassAnnotations($class);
 
             if ($annotations === []) {
                 continue;
             }
 
-            yield $class->getFileName() => [$definition, $annotations];
+            $fileName = $class->getFileName();
+            assert(is_string($fileName));
+
+            yield $fileName => [$definition, $annotations];
         }
     }
 
-    /**
-     * @param Mapping\Annotation[] $annotations
-     */
+    /** @param Mapping\Annotation[] $annotations */
     private function appendTags(Definition $definition, array $annotations): void
     {
         foreach ($annotations as $annotation) {
@@ -98,9 +106,7 @@ final class ExpandTags implements CompilerPassInterface
         }
     }
 
-    /**
-     * @return string[]
-     */
+    /** @return array<string, string> */
     private function createAttributes(Mapping\Annotation $annotation): array
     {
         $attributes = (array) $annotation;
